requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
streamlit==1.29.0
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
